backend/requirements_deploy.txt
# -------------------------------------------------------
# Production / deployment requirements (Render free tier)
# Install with:  pip install -r requirements_deploy.txt
#
# NOTE: sentence-transformers and torch are NOT included.
# The RAG pipeline loads precomputed embeddings from
# models/embeddings.pkl into an ephemeral ChromaDB client,
# so PyTorch is never needed at runtime.
# -------------------------------------------------------

# Web framework + server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Config
python-dotenv>=1.0.0

# ML inference (model + vectorizer are loaded via joblib)
scikit-learn>=1.4.0
joblib>=1.3.0

# Data handling
pandas>=2.1.0
numpy>=1.26.0

# NLP
nltk>=3.9.0

# RAG — ephemeral in-memory vector DB (ONNX-based, no torch)
chromadb>=0.5.0

# LangGraph agent orchestration
langgraph>=0.1.0
langchain-core>=0.2.0

# LLM providers
requests>=2.31.0
langchain-nvidia-ai-endpoints>=0.1.0

# Misc
tqdm>=4.66.0
typing_extensions>=4.7.0
